Move `NetStream` to a new mod and clean up mtsender net code
