+++ lib/grammers-mtsender/src/lib.rs
@@ -9,6 +9,7 @@
 #![deny(unsafe_code)]
 
 mod errors;
+mod net;
 mod reconnection;
 pub mod utils;
 
@@ -23,6 +24,8 @@ use grammers_mtproto::transport::{self, Transport};
 use grammers_mtproto::{authentication, MsgId};
 use grammers_tl_types::{self as tl, Deserializable, RemoteCall};
 use log::{debug, error, info, trace, warn};
+use net::NetStream;
+pub use net::ServerAddr;
 use std::io;
 use std::io::Error;
 use std::ops::ControlFlow;
@@ -31,23 +34,11 @@ use std::sync::atomic::{AtomicI64, Ordering};
 use std::time::Duration;
 use tl::Serializable;
 use tokio::io::{AsyncReadExt, AsyncWriteExt};
-use tokio::net::tcp::{ReadHalf, WriteHalf};
-use tokio::net::TcpStream;
 use tokio::sync::mpsc;
 use tokio::sync::oneshot;
 use tokio::sync::oneshot::error::TryRecvError;
 use utils::{sleep, sleep_until};
 use web_time::{Instant, SystemTime};
-
-#[cfg(feature = "proxy")]
-use {
-    hickory_resolver::config::{ResolverConfig, ResolverOpts},
-    hickory_resolver::AsyncResolver,
-    std::io::ErrorKind,
-    std::net::{IpAddr, SocketAddr},
-    tokio_socks::tcp::Socks5Stream,
-    url::Host,
-};
 
 /// The maximum data that we're willing to send or receive at once.
 ///
@@ -96,31 +87,12 @@ pub(crate) fn generate_random_id() -> i64 {
     LAST_ID.fetch_add(1, Ordering::SeqCst)
 }
 
-pub enum NetStream {
-    Tcp(TcpStream),
-    #[cfg(feature = "proxy")]
-    ProxySocks5(Socks5Stream<TcpStream>),
-}
-
-impl NetStream {
-    fn split(&mut self) -> (ReadHalf, WriteHalf) {
-        match self {
-            Self::Tcp(stream) => stream.split(),
-            #[cfg(feature = "proxy")]
-            Self::ProxySocks5(stream) => stream.split(),
-        }
-    }
-}
-
-// Manages enqueuing requests, matching them to their response, and IO.
-
+/// Manages enqueuing requests, matching them to their response, and IO.
 pub struct Sender<T: Transport, M: Mtp> {
     stream: NetStream,
     transport: T,
     mtp: M,
-    addr: std::net::SocketAddr,
-    #[cfg(feature = "proxy")]
-    proxy_url: Option<String>,
+    addr: ServerAddr,
     requests: Vec<Request>,
     request_rx: mpsc::UnboundedReceiver<Request>,
     next_ping: Instant,
@@ -190,13 +162,13 @@ impl Enqueuer {
 }
 
 impl<T: Transport, M: Mtp> Sender<T, M> {
-    async fn connect<'a>(
+    async fn connect(
         transport: T,
         mtp: M,
-        addr: std::net::SocketAddr,
+        addr: ServerAddr,
         reconnection_policy: &'static dyn ReconnectionPolicy,
     ) -> Result<(Self, Enqueuer), io::Error> {
-        let stream = connect_stream(&addr).await?;
+        let stream = NetStream::connect(&addr).await?;
         let (tx, rx) = mpsc::unbounded_channel();
         Ok((
             Self {
@@ -204,41 +176,6 @@ impl<T: Transport, M: Mtp> Sender<T, M> {
                 transport,
                 mtp,
                 addr,
-                #[cfg(feature = "proxy")]
-                proxy_url: None,
-                requests: vec![],
-                request_rx: rx,
-                next_ping: Instant::now() + PING_DELAY,
-                reconnection_policy,
-
-                read_buffer: vec![0; MAXIMUM_DATA],
-                read_tail: 0,
-                write_buffer: DequeBuffer::with_capacity(MAXIMUM_DATA, LEADING_BUFFER_SPACE),
-                write_head: 0,
-            },
-            Enqueuer(tx),
-        ))
-    }
-
-    #[cfg(feature = "proxy")]
-    async fn connect_via_proxy<'a>(
-        transport: T,
-        mtp: M,
-        addr: SocketAddr,
-        proxy_url: &str,
-        reconnection_policy: &'static dyn ReconnectionPolicy,
-    ) -> Result<(Self, Enqueuer), io::Error> {
-        info!("connecting...");
-
-        let stream = connect_proxy_stream(&addr, proxy_url).await?;
-        let (tx, rx) = mpsc::unbounded_channel();
-        Ok((
-            Self {
-                stream,
-                transport,
-                mtp,
-                addr,
-                proxy_url: Some(proxy_url.to_string()),
                 requests: vec![],
                 request_rx: rx,
                 next_ping: Instant::now() + PING_DELAY,
@@ -366,17 +303,7 @@ impl<T: Transport, M: Mtp> Sender<T, M> {
     async fn try_connect(&mut self) -> Result<(), Error> {
         let mut attempts = 0;
         loop {
-            #[cfg(feature = "proxy")]
-            let res = if self.proxy_url.is_some() {
-                connect_proxy_stream(&self.addr, self.proxy_url.as_ref().unwrap()).await
-            } else {
-                connect_stream(&self.addr).await
-            };
-
-            #[cfg(not(feature = "proxy"))]
-            let res = connect_stream(&self.addr).await;
-
-            match res {
+            match NetStream::connect(&self.addr).await {
                 Ok(result) => {
                     log::info!(
                         "auto-reconnect success after {} failed attempt(s)",
@@ -782,85 +709,11 @@ impl<T: Transport> Sender<T, mtp::Encrypted> {
 
 pub async fn connect<T: Transport>(
     transport: T,
-    addr: std::net::SocketAddr,
+    addr: ServerAddr,
     rc_policy: &'static dyn ReconnectionPolicy,
 ) -> Result<(Sender<T, mtp::Encrypted>, Enqueuer), AuthorizationError> {
     let (sender, enqueuer) = Sender::connect(transport, mtp::Plain::new(), addr, rc_policy).await?;
     generate_auth_key(sender, enqueuer).await
-}
-
-#[cfg(feature = "proxy")]
-pub async fn connect_via_proxy<'a, T: Transport>(
-    transport: T,
-    addr: std::net::SocketAddr,
-    proxy_url: &str,
-    rc_policy: &'static dyn ReconnectionPolicy,
-) -> Result<(Sender<T, mtp::Encrypted>, Enqueuer), AuthorizationError> {
-    let (sender, enqueuer) =
-        Sender::connect_via_proxy(transport, mtp::Plain::new(), addr, proxy_url, rc_policy).await?;
-    generate_auth_key(sender, enqueuer).await
-}
-
-async fn connect_stream(addr: &std::net::SocketAddr) -> Result<NetStream, std::io::Error> {
-    info!("connecting...");
-    Ok(NetStream::Tcp(TcpStream::connect(addr).await?))
-}
-
-#[cfg(feature = "proxy")]
-async fn connect_proxy_stream(
-    addr: &SocketAddr,
-    proxy_url: &str,
-) -> Result<NetStream, std::io::Error> {
-    let proxy =
-        url::Url::parse(proxy_url).map_err(|err| io::Error::new(ErrorKind::InvalidData, err))?;
-    let scheme = proxy.scheme();
-    let host = proxy.host().ok_or(io::Error::new(
-        ErrorKind::NotFound,
-        format!("proxy host is missing from url: {}", proxy_url),
-    ))?;
-    let port = proxy.port().ok_or(io::Error::new(
-        ErrorKind::NotFound,
-        format!("proxy port is missing from url: {}", proxy_url),
-    ))?;
-    let username = proxy.username();
-    let password = proxy.password().unwrap_or("");
-    let socks_addr = match host {
-        Host::Domain(domain) => {
-            let resolver = AsyncResolver::tokio(ResolverConfig::default(), ResolverOpts::default());
-            let response = resolver.lookup_ip(domain).await?;
-            let socks_ip_addr = response.into_iter().next().ok_or(io::Error::new(
-                ErrorKind::NotFound,
-                format!("proxy host did not return any ip address: {}", domain),
-            ))?;
-            SocketAddr::new(socks_ip_addr, port)
-        }
-        Host::Ipv4(v4) => SocketAddr::new(IpAddr::from(v4), port),
-        Host::Ipv6(v6) => SocketAddr::new(IpAddr::from(v6), port),
-    };
-
-    match scheme {
-        "socks5" => {
-            if username.is_empty() {
-                Ok(NetStream::ProxySocks5(
-                    tokio_socks::tcp::Socks5Stream::connect(socks_addr, addr)
-                        .await
-                        .map_err(|err| io::Error::new(ErrorKind::ConnectionAborted, err))?,
-                ))
-            } else {
-                Ok(NetStream::ProxySocks5(
-                    tokio_socks::tcp::Socks5Stream::connect_with_password(
-                        socks_addr, addr, username, password,
-                    )
-                    .await
-                    .map_err(|err| io::Error::new(ErrorKind::ConnectionAborted, err))?,
-                ))
-            }
-        }
-        scheme => Err(io::Error::new(
-            ErrorKind::ConnectionAborted,
-            format!("proxy scheme not supported: {}", scheme),
-        )),
-    }
 }
 
 pub async fn generate_auth_key<T: Transport>(
@@ -903,8 +756,6 @@ pub async fn generate_auth_key<T: Transport>(
             write_buffer: sender.write_buffer,
             write_head: sender.write_head,
             addr: sender.addr,
-            #[cfg(feature = "proxy")]
-            proxy_url: sender.proxy_url,
             reconnection_policy: sender.reconnection_policy,
         },
         enqueuer,
@@ -913,7 +764,7 @@ pub async fn generate_auth_key<T: Transport>(
 
 pub async fn connect_with_auth<T: Transport>(
     transport: T,
-    addr: std::net::SocketAddr,
+    addr: ServerAddr,
     auth_key: [u8; 256],
     rc_policy: &'static dyn ReconnectionPolicy,
 ) -> Result<(Sender<T, mtp::Encrypted>, Enqueuer), io::Error> {
@@ -921,24 +772,6 @@ pub async fn connect_with_auth<T: Transport>(
         transport,
         mtp::Encrypted::build().finish(auth_key),
         addr,
-        rc_policy,
-    )
-    .await
-}
-
-#[cfg(feature = "proxy")]
-pub async fn connect_via_proxy_with_auth<'a, T: Transport>(
-    transport: T,
-    addr: std::net::SocketAddr,
-    auth_key: [u8; 256],
-    proxy_url: &str,
-    rc_policy: &'static dyn ReconnectionPolicy,
-) -> Result<(Sender<T, mtp::Encrypted>, Enqueuer), io::Error> {
-    Sender::connect_via_proxy(
-        transport,
-        mtp::Encrypted::build().finish(auth_key),
-        addr,
-        proxy_url,
         rc_policy,
     )
     .await
